added make an appointment botton and create makeAnAppointment screen
