+++ pubspec.yaml
@@ -79,6 +79,7 @@ dependencies:
   google_maps_flutter: ^2.2.8
   location: ^7.0.0
   geolocator: ^12.0.0
+  flutter_map: ^7.0.2
   # flutter_plugin_android_lifecycle: ^2.0.21
 
 dev_dependencies:
